+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Currently this project is in alpha stage.
 ### Added
 
 - JSON schema for server configuration
+- Setting `boomack.client.fileSrc.enable`
+- Setting `boomack.client.fileSrc.hosts`
 
 ## 0.1.7